Fix migrations, use simple `get_client` function, migrate AuthorizationsConfig
